refactor: split code for better modularity and rename variables for clarity
