+++ src/features/user/MFV2/ForestProgress/ForestProgress.module.scss
@@ -1,6 +1,6 @@
 @import '../../../../theme/theme';
 
-.targetMainContainer {
+.progressSection {
   display: flex;
   gap: 10px;
   width: 100%;
