+++ src/features/user/MFV2/ForestProgress/microComponents/GiftPercentageBar.tsx
@@ -1,5 +1,5 @@
 import progressBarStyle from '../ForestProgress.module.scss';
-import { BarsProps } from '../ForestProgressItem';
+import { BarsProps } from './Bars';
 
 const GiftPercentageBar = ({
   personalPercentage,
